double instead of float in graphics API

+++ src/comp127graphics/Arc.java
@@ -89,12 +89,12 @@ public class Arc extends GraphicsObject implements Strokable {
         setStroked(true);
     }
 
-    public float getStrokeWidth() {
+    public double getStrokeWidth() {
         return stroke.getLineWidth();
     }
 
-    public void setStrokeWidth(float width) {
-        stroke = new BasicStroke(width);
+    public void setStrokeWidth(double width) {
+        stroke = new BasicStroke((float) width);
         changed();
     }
